Make FileSystemService and ImageService reactive (deferred to separate scheduler)

+++ reactive-albums-api/src/main/kotlin/nl/juraji/reactive/albums/configuration/PicturesAggregateConfiguration.kt
@@ -19,8 +19,8 @@ data class PicturesAggregateConfiguration(
         val thumbnailLocation: String,
         val thumbnailType: String,
         val hashingSampleSize: Int,
-        val autoCropTolerance: Float,
-        val duplicateSimilarity: Float,
+        val autoCropTolerance: Double,
+        val duplicateSimilarity: Double,
 ) {
     val thumbnailMimeType: MediaType = MediaType.parseMediaType(thumbnailType)
     val thumbnailPath: Path = Paths.get(thumbnailLocation)
